Move integer const-folding tests to a more suitable place.
