+++ hugr/src/std_extensions/arithmetic/int_ops/const_fold.rs
@@ -1199,3 +1199,6 @@ pub(super) fn set_fold(op: &IntOpDef, def: &mut OpDef) {
         },
     });
 }
+
+#[cfg(test)]
+mod test;
